jumping and falling fixes
Jumping mechanics allowed for the player to jump once off the ground, and once again in air. Changed the location of the reset, and jump mechanic things to end step to fix.

Also added a fixed, and working slowed fall when on ledge mechanic. When falling next to a ledge, and the player goes against it, the players falling speed is reduced.